refactor: 로또 비교 기능 LottoGame → PurchasedLotto로 이동
- 파라미터 줄이기
